Fix subtitle loading with network URIs
When a user switched to another app and switches back, if the activity is still
alive, the previously found subtitles (by `searchSubtitles`) won't be loaded because:
1. The player instance is already released in `Activity#onStop`, and `searchSubtitle`
are only called when creating the activity or when it receives new intents. But in this
situation only `Activity#onStart` would be called.
2. `Activity#onStart` calls `initializePlayer`, which tries to load the previously
used subtitleUri, but `Utils.fileExists` does not handle HTTP URIs.

This adds a check in `initializePlayer` to try loading subtitles with network URIs.